am33xx: Re-enable SW levelling for DDR2
The changes in 9498ca5 were not intended for DDR2 boards, only DDR3.
Update emif_sdram_type to take a sdram_config value to check against.
This lets us pass in the value we would use to configure, when we have
not yet configured the board yet.  In other cases update the call to be
as functional as before and check an already programmed value in.

Tested-by: Yan Liu <yan-liu@ti.com>
Signed-off-by: Tom Rini <trini@ti.com>