usb: gadget: ether: defer setting maxpacket till ->setup()
This is actually needed to get rndis enumerate with DWC3 device.
While the actual fix should be to make ethernet gadget use the composite
driver, defering to set the maxpacket till ->setup can be used as a
temporary fix.

This is similar to the patch by Sebasetian in mainline kernel
commit 765f5b830e547229bb752e7b232ee83e2b3d49d5
Author: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Date:   Thu Jun 23 14:26:11 2011 +0200

    usb: gadget: defer setting maxpacket till ->setup()

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>