usb: gadget: ether: initialize and cleanup gadget from ether gadget
Initializing the dwc3 gadget driver and not disabling it results in
dwc3 being in a state which prevents system from suspending while doing
a PM in the linux kernel. By moving the initialization function
to the ether gadget gives a place where cleanup function for
dwc3 driver can be invoked.

Reported-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>