HACK: ARM: DRA7: Configure DSP & IVA voltage domains for OPP_HIGH
The current bootloaders has all the Voltage domains setup and
configured for OPP_NOM. Update the voltage domain data on DSP
and IVA Voltage domains to be configured for OPP_HIGH. This is
being done to meet the performance needs of 1080p MultiMedia
usecases and other DSP usecases. This change is done for all
the existing DRA7xx/AM57xx platforms.

The OPP_HIGH voltage settings are done in u-boot and the clock
configuration done in kernel because of the following reasons:
1. SoC definition constraints us to NOT to do dynamic voltage
   scaling ever after the initial avs0 setting in bootloader
  - so the voltage must be set in bootloader.
2. The voltage level must be set even if the IP blocks like
   SGX/DSP are unused.
3. The IVA and DSP DPLLs are not essential for u-boot
   functionality, and similar DPLL clock configuration code
   has been cleaned up in v2014.10 kernel. See commit,
   02c4153 ("ARM: OMAP4/5: Remove dead code against
   CONFIG_SYS_CLOCKS_ENABLE_ALL").

The non-essential DPLLs are configured within the kernel during
the clock init step when parsing the device tree and creating
the clock devices. This approach meets both the u-boot and kernel
needs.

The GPU voltage domain is also updated to OPP_HIGH on platforms
with ganged voltage rails supplying either of the IVA or DSP
voltage domains. Following is the complete list of voltage
domains updated:
   AM57xx EVM / Beagle-x15 : IVA, DSPEVE & GPU (SMPS45)
   AM571x / AM572x IDK     : IVA, DSPEVE & GPU (SMPS45)
   DRA72x EVM              : IVA, DSPEVE & GPU (SMPS3)
   DRA74x EVM              : IVA (SMPS8) and DSPEVE (SMPS45)

Signed-off-by: Suman Anna <s-anna@ti.com>