configs: ti_armv7_keystone2: Get pmmc binary from {bootdir} in UBI
The PMMC firmware binary is currently fetched from different locations
between the MMC boot and UBI boot modes for K2G SoC. Fix up the
get_pmmc_ubi env variable so that the PMMC firmware binary location
in UBI mode matches that of MMC boot. This allows the file system
to place the PMMC binary at the same place as the kernel, DT blob
and boot monitor binaries, and make it identical for MMC and UBI
boot modes.

Fixes: 123e72bd60a9 ("ARM: keystone: K2G: Add PMMC remoteproc device")
Signed-off-by: Suman Anna <s-anna@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>
Acked-by: Denys Dmytriyenko <denys@ti.com>