ti: AM437x: Add gpi2c_init before attempting to scale MPU voltage rail
With the new board detection logic introduced in e3ad3a2ca4ec ("ti:
AM437x: Use generic EEPROM detection logic") the board eeprom is read
once and then cached in sram. In the case of a warm reset on am437x,
the context of sram is maintained though the reset so read_eeprom is
able to fall back to this data, however because of the fallback the
internal gpi2c_init call is skipped which previously was configuring
the i2c bus before the PMIC voltage change functions were called
for the MPU voltage rail.

Because of this, PMIC functions at this early stage of boot fail during
a warm reset because the i2c bus has not been initialized. Add an
explicit call to gpi2c_init before attempting to scale the voltage rails
so that we are always able to communicate with the PMIC. This path that
fails is unique to am437x board configuration.

Acked-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>