board: ti: am57xx: update RGMII1 io-delay values for 2.0 silicon
The 1.3a rev of the AM572x-IDK board has 2.0 silicon on it which
requires an update to the io-delay values for cpsw ethernet port 1.
The RGMII1 IO Delay values are 2.0 estimated values. These
values will need to be verified once device characterization is
complete. Characterization of the 2.0 device will be a significant
time from now. Without this update for 2.0 silicon the RGMII1 will
not work.

These values for RGMII1 will work on 1.1 and 2.0 silicon. The
datasheet used is SPRS953P from November 2015.

Signed-off-by: Schuyler Patton <spatton@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>