configs: k2g_evm/keyston2: Fix pmmc binary name update
When we use the following in bootargs:
v1=abc
v2=123-${v1}
echo ${v2}
we get 123-${v1}
This was discussed in https://patchwork.ozlabs.org/patch/552864/
and Wolfgang disagrees that we should do recursive variable parsing.

So, fix it up to parse during run command itself. Kind of a little
convoluted, but anyways..

Fixes: c83bd9820e9a ("configs: k2g_evm/keystone2: Rename pmmc binary to be generic")
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>