ti: AM437x: Fix gpi2c_init call protection
For AM437x, gpi2c_init relies on a static local variable 'init' to
ensure the i2c initialization routines only get called once from
bdbd76b04c15 ("ti: AM437x: Add gpi2c_init before attempting to
scale MPU voltage rail"). However, this function is called very early,
before BSS is available, so we cannot rely on the value of 'init' to be
zero on first access, especially after a warm reset. Because of this it
is possible that gpi2c_init never initializes i2c at all, leading to
MPU voltage scaling not happening which can cause boot to fail.

Also, we need to initialize the variable to a non-zero value so that the
compiler does not 'optimize' and place the value in bss anyway.

In order to avoid this, let's rename the variable to 'first_time' to
completely avoid any confusion, initialize it to true and then set the
variable to false after we run once so that we don't re-initialize i2c
when the function is called again.

Acked-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>