board: am57xx: Fix sysinfo string length
sizeof a pointer will always return the size of the data type it is pointing
to. Board detection logic for am57xx prints the board name into sysinfo
with size as pointer size. Because of this only 3 letters are printed: "Boa".
So, pass the maximum size for printing. Also print the Revision of the board
along with it.

Fixes: ab0e3372ff2e ("board: ti: AM57xx: Add detection logic for AM57xx-evm")
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>