board: ti: am57xx: Switch HDMI muxing on SR 2.0 for X15/GPEVM
HDMI_LS_OE from GPIO6_28(Y9) to GPIO2_30 (AG8) to avoid to avoid a
1.8V GPIO toggling a 3.3V input when the SD card in HS 1.8V mode. This
requires a minor change in our configuration for ES1.1 Vs ES2.0 setup.
We split the HDMI LS_OE configuration based on the ES of the silicon
and by default configure the pinmux for MMC1_SDWP as driver_off.

Signed-off-by: Nishanth Menon <nm@ti.com>