ti: AM437x: Fix HS-device variant related board init
With the introduction of support for the high-security (HS) variant of
the AM437x GPEVM a function to identify this EVM was introduced in commit
894e719aadca2f ("ti: AM43xx: board: Detect AM43xx HS EVM"). What was missing
however was to update all AM437x GPEVM related board init code to also get
invoked in case of using this newly added EVM. In order to arrive at an
easy to maintain solution moving forward in addition to fixing the issue
at hand, introduce a board detection function that does not distinguish
between GPEVM and HS EVM for cases where this is not needed, and utilize
this function in all applicable places.

The missing invocation of board init code manifests itself in significantly
reduced memory bandwidth observable during benchmarking but also
generating issues with display output.

Fixes: 894e719aadca2f ("ti: AM43xx: board: Detect AM43xx HS EVM")
Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>