ARM: OMAP5/DRA7: Get rid of control_std_fuse_opp_vdd_mpu_2
commit 2d9d057be6aaa410650b7ceecfcdb0aca4c24e31 upstream.

This information is already available under vcores->volts.efuse.reg.
There is no reason for duplicating the information since AVS Class 0
definitions are common for OMAP5 and DRA7 and defined with
STD_FUSE_OPP_* macros. This allows a central location of defining
the ABB and voltage definitions especially since they are reused.

This also makes it simpler to prevent mistakes involved when changing
the boot OPP for the device.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Gabriel Huau <ghuau@ti.com>