mmc: trust MMC CSD to report the correct MMC mode
commit 89a8ca7cef676a2b1ba43a3e7265872d5cc3eea2 branch ti-u-boot-2016.05

Switch the MMC card to "high speed" if the card capability indicates
that it supports high speed instead of directly switching to high
speed mode to see if the card supports high speed.

Reading the CSD again to see if the HS_TIMING bit is set shouldn't
be required. However that doesn't harm and hence left it as it is.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>