drivers: mmc: omap_hsmmc: add support to set default io voltage
commit 064d902822834ba511346b1b5c64411d6dc5360a branch ti-u-boot-2016.05

"ti,dual-volt" is used in linux kernel to set the voltage capabilities.
For host controller dt nodes that doesn't have "ti,dual-volt",
it's assumed 1.8v is the io voltage. This is not always true (like in
the case of beagle-x15 where the io lines are connected to 3.3v).
So add a few #define to define the io volatage, and the dual-voltage
capability.

If CONFIG_NO_HSMMC_DUALT_VOLT is not set, then dual-voltage support is
active.
If CONFIG_HSMMCx_IOV (where x is 1, 2, or 3) is not set, the default io
voltage of 1.8v will be used. However if dual-voltage is active, the
default io voltage will be set to 3v

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/mmc/omap_hsmmc.c