ARM: OMAP5/DRA7: Enable iodelay recalibration to be done from uboot
commit e376b80c4f9388f69d529a61398bfae7cf9b74d9 branch ti-u-boot-2016.05

Add a new API to perform iodelay recalibration without isolate
io to be used in uboot.

The data manual of J6/J6 Eco recommends to set different IODELAY values
depending on the mode in which the MMC/SD is enumerated in order to
ensure IO timings are met. The MMC driver can use the new API to
set the IO delay values depending on the MMC mode.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	arch/arm/include/asm/arch-omap5/dra7xx_iodelay.h
	include/configs/am57xx_evm.h