drivers: mmc: omap_hsmmc: Add support to set IODELAY values
commit c791e6c2d5742e9826452a03d8df777ca6ad2466 branch ti-u-boot-2016.05

The data manual of J6/J6 Eco recommends to set different IODELAY values
depending on the mode in which the MMC/SD is enumerated in order to
ensure IO timings are met.

Add support to get mux values and iodelay values from platfrom specific
code and set these depending on the enumerated MMC mode.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/mmc/omap_hsmmc.c

Conflicts:
	drivers/mmc/omap_hsmmc.c