mmc: add support to switch the SD card to UHS mode
Add support to switch the voltage to 1.8V required for UHS mode. And
add support to switch the SD card to UHS modes (SDR104/DDR50/SDR50/SDR25/
SDR12). Power cycle support is not added and hence if voltage switching
fails, the card should be replaced and the board should be reset.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>