ARM: dts: dra7xx: am57xx: Update speed modes for MMC1
Add capabilities in MMC1 dt node to indicate it supports SDR104,
DDR50, SDR50, SDR25, SDR12 and HS modes. Also add iodelay values for
the different MMC modes. Since power cycle can't be supported
in am57xx-idk and am57xx-beagle-x15 because of the way the vdd
line is connected, the pinctrl values for UHS modes
in these boards are not populated.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>