drivers: net: keystone_net: add gbe serdes lane and second serdes configurations
SerDes configuration is divided mainly in two parts, namely,
a common configuration which is done regardless a SerDes lane
is actually used or not, and a lane specific configuration
which is done when a lane is actually used

This patch adds the detailed lane specific configurations as
is needed in the second part of a serdes configurations.

This patch also adds a separate data structure for the second
serdes found on K2E and K2L to support the flexibility of
configuring each serdes with different parameters.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>