HACK: ti: AM335x: assume am335x_evm with HS to reduce SPL size
To reduce the size of SPL on AM335x HS devices we hardcode the board
detection code to detect am335x_evm when HS build is enabled. This
is the only HS board we currently support and this lets the compiler
drop all detection and initialization code for other boards.

A better solution may be to split these boards out into different
targets, instead of now where they all claim to be
CONFIG_TARGET_AM335X_EVM's in their defconfigs.

Signed-off-by: Andrew F. Davis <afd@ti.com>