drivers: net: keystone_net: fix errror in serdes signal detect
commit b15c6facd608 ("drivers: net: keystone_net: add gbe serdes
lane and second serdes configurations") added the lane specific
enablement before the phy is enabled. This would cause a signal
detect issue in the gbe serdes. SerDes lane rx should be enabled
only after phy is connected, for otherwise SerDes lane will not
be able to detect the signal. This patch moves the calling of
serdes lane enable rx to after phy_connect is called.

Fixes: b15c6facd608 ("drivers: net: keystone_net: add gbe serdes
lane and second serdes configurations")

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Tested-by: Suman Anna <s-anna@ti.com>