drivers: rapidio: keystone: fix rapidio boot failure at low temperature
RapidIO boot can fail when the DUTs are in a low temperature (0 to -40C)
environment. This patch updates the RapidIO SerDes initialization
configuration to fix the problem. Also added information on which
CSL version are the changes aligned to in the Revision section in
header for easy maintenance in future debugging.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Arun Uppuleti <auppuleti@ti.com>