spi: cadence_qspi_apb: Improve indirect write transaction
If the write transaction size(write_bytes) is not a multiple of word
length, then issue word length writes till the we reach the dangling
bytes. On the final write, issue byte by byte write to complete the
transaction. This marginally improves write throughput when performing
random sized writes to the flash.

Signed-off-by: Vignesh R <vigneshr@ti.com>