TI: Rework SRAM definitions and maximums for dra7xx platform
commit fa2f81b06f666710c756d25297d7a9ca48c65935 upstream.

On all TI platforms the ROM defines a "downloaded image" area at or near
the start of SRAM which is followed by a reserved area.  As it is at
best bad form and at worst possibly harmful in corner cases to write in
this reserved area, we stop doing that by adding in the define
NON_SECURE_SRAM_IMG_END to say where the end of the downloaded image
area is and make SRAM_SCRATCH_SPACE_ADDR be one kilobyte before this.
At current we define the end of scratch space at 0x228 bytes past the
start of scratch space this this gives us a lot of room to grow.  As
these scratch uses are non-optional today, all targets are modified to
respect this boundary.

Signed-off-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>