fs: fat: use debug() for buffer misalignment warnings
commit 62605654f2cda82c1f1f1c1045726edde746d782 branch ti-u-boot-2017.01

functions to get/set the data from/to the block device require buffers
aligned on the cache lines. When the buffers are not aligned, a bounce
buffer is used and a warning message is output. This has been introduced
in commits cc63b25e and 8133f43d1.
As this is only a warning, let's move to using debug().

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>