fs: fat: use debug() for buffer misalignment warnings
functions to get/set the data from/to the block device require buffers
aligned on the cache lines. When the buffers are not aligned, a bounce
buffer is used and a warning message is output. This has been introduced
in commits cc63b25e and 8133f43d1.
As this is only a warning, let's use debug() instead of printf().

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>