HACK: config: am335x_evm: undef config options for HS
On AM335x HS devices we do not have enough SRAM to load an SPL image
that can read from every possible media type, so we disable media types
based on defconfig.

As the config header to Kconfig transition is still in flux upstream,
for now we use ifdefs to undef options, a better solution will be to do
this in defconfig when that is available.

Remove SPL EFI partition support for HS AM335x EVMs.

Signed-off-by: Andrew F. Davis <afd@ti.com>