mmc: Add a new _ios_ parameter to set timing
Add a new _ios_ parameter to set the timing specification
used. It is now set to HS/DDR/HS200 mode and it can be used
to set other UHS modes once support for UHS mode is added.
The host controller driver can use this to perform any
mode specific settings.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>