mmc: add a library function to send tuning command
HS200/SDR104 requires tuning command to be sent to the card.
Add a simple function to send tuning command and to read and
compare the received data with the tuning block pattern.
This function can be used by platform driver to perform DLL
tuning.
This patch is similar to
commit 996903de92f0c7a32d8c83f37d7ebcea0def8660 ("mmc: core: add
core-level function for sending tuning commands") added in linux
kernel.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>