ARM: k2g: Run pinmux configuration again when board detection is complete
Initially when the board is booted up it doesn't know which board it
is. Therefore, initially only configure generic pinmuxing that is
universal across all of TI's 66AK2GX evms. Once the EEPROM is read the
system can finally determine the correct pinmuxing to use so perform
pinmuxing again with the additional pins.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>