mmc: dm: spl: remove special handling of the SPL case
If we handle the probing the same way in the spl and in u-boot, we can
simplify the code in spl_mmc. And the controller mapping will be
the same in the spl and in u-boot, making it much easier to have the
environment on a MMC device that is not the boot device.

Tested-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>