ARM: k2g: Add alternative runtime board detection support
K2G GP boards do not currently have their EEPROM programmed. This is
problematic since that is the primary way to do runtime board detection.
However, an alternative approach can be used which is probing i2c ics that
are only found on K2G GP evm. Once we know we are on the K2G GP we then
fake the EEPROM contents to allow the remaining code to work as normal.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>