configs: am335x_evm: Enable tiny printf in SPL
am335x_evm SPL is very close to its limit in SRAM space.
SRAM size might be overflowed when building with "O=" option
with a very large directory name and the same is a sibling
of source tree. This will give the absolute path when __FILE_
is used.
Switch to use tiny printf to reclaim some size.

Reported-by: Denys Dmytriyenko <denys@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>