configs: davinci: omapl138_lcdk: add random eth address support
Any TFTP or DHCP boot on the Davinci OMAP-L138 LCDK board requires
that the ethaddr variable be defined. There are no e-fuses to store
the ethernet mac address for this platform, and neither is a mac
address reserved in any format. So enable random mac address support
so that networking boot can be supported.

Signed-off-by: Suman Anna <s-anna@ti.com>