boot: fdt: Perform arch_fixup_fdt() on the given device tree for falcon boot
commit 6e7585bb64b12 upstream

In single stage bootmode or falcon boot mode, the SPL shall update the
device tree that we load with the normal fixups done via
arch_fixup_fdt(), when possible (ie we have enough information in this
restricted environment to be able to do that still).  This will include
for example updating them memory nodes.

Signed-off-by: Ravi Babu <ravibabu@ti.com>
[trini: Reword commit message]
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>