drivers: omap_hsmmc: If DM_MMC is not used, get the iodelays and the pinmux from the platform code
based on commit e762451dbcc2 ("drivers: omap_hsmmc: In SPL get the iodelays
and the pinmux from the platform code") branch ti-u-boot-2017.01

The iodelays and pinmux information is platform specific. Since the SPL
can embed only one dts, getting this information from it would mean to have
one SPL binary for each platform.
In order to keep the ability to run the same SPL binary on several
platforms, the iodelays and pinmux configurations used by the SPL are
provided by the platform code.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>