xyz-modem: Fix timeout loop waiting with WATCHDOG
Commit 2c77c0d6524eb ("xyz-modem: Change getc timeout loop waiting")
fixes the loop delay when using a hw watchdog, assuming that Watchdog
kicking is taken care of by getc(). But watchdog kicking in getc happens
only when a request for retry is occurred. This request might not happen
all the time. So, kick the watchdog before loop waiting, instead of
relying on other functions.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>