arm: dra76x: cmd_ddr: Use ecc_test env variable for 2-bit error
On occurrence of a 2-bit ECC error, system needs to be reset.
In order to test this feature in kernel, u-boot should inject
2-bit error and not reset the device. When EDAC driver gets probed
in kernel, it sees a 2-bit error and triggers a kernel panic.
So, introduce a ecc_test env variable, which when set u-boot will
not reset the device on a 2-bit error.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>