mmc: omap_hsmmc: Workaround errata regarding SDR104/HS200 tuning failures (i929)
There is a silicon errata in DRA75x, DRA72x and OMAP5 platforms which
may result in unexpected tuning pattern errors. A small failure band
may be present in the tuning range which may be missed by the curent
algorithm. Furthermore, the failure bands vary with temperature leading
to different optimum tuning values for different temperatures.

As suggested in the silicon errata, tuning should be done in two stages.
In stage 1, assign the optimum tuning ratio in the maximum pass window
for the current temperature. In stage 2, if the chosen value is close
to the small failure band, move away from it in the appropriate direction.

Therefore, modify SDR104/HS200 DLL tuning algorithm as a workaround.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>