env: mmc/fat/ext4: Initialize mmc in environment only in SPL
Calls to mmc_initialize() in u-boot before relocation lead to mmc
not being detected after relocation as the pointers are not updated
across relocation. These calls to mmc_initialize() were added to
initialize mmc when environment is required for SPL. In the case
of u-boot, mmc devices are initialized and environment is obtained
after relocation anyway. Therefore limit mmc_initialize() to SPL.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>