configs: k2g_evm: fix ethernet phy configuration
Configure k2g EVM for the exact PHY part that
is present on the board. This makes U-Boot apply
configurations needed for this PHY like centering
the FLP timing, and applying the necessary skews
for RGMII tx, rx, clock and control lines.

This helps solve an intermittent issue (5 of 1000
times) where U-Boot is unable to establish ethernet
link over gigabit network.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>