configs: am57xx_evm: fix ethernet phy configuration
Configure AM57xx EVMs for the exact PHY part that is
present on the various boards. This makes U-Boot apply
configurations needed for this PHY like centering the
FLP timing, and applying the necessary skews for RGMII
tx, rx, clock and control lines.

For configurations to take effect, DM_ETH needs to be
enabled. Do that too.

Tested on AM572x GP EVM, AM574x IDK and AM571x IDK.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>