mmc: omap_hsmmc: Handle command and data errors after response
It is recommended that SRC be asserted when command error bits are
asserted and SRD be asserted when data error bits are asserted after a
command is sent. However, during tuning SRD must never be asserted as
tuning data will be lost.
Therefore, add support for handling command and data error bits after a
command is sent except when tuning is happening.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>

Reviewed-by: Kishon Vijay Abraham I <kishon@ti.com>