ARM: dts: am574x-idk: Remove UHS modes in am574x-idk
If an SD card enumerates in a UHS mode and undergoes a voltage
switch to 1.8V in U-boot, the S18A bit in OCR of the SD card will be
set to 0 indicating that it has switched to 1.8V. To set this bit
back to 1, a power cycle of the card is required. As am574x-idk
doesn't have power cycle support, this bit will remain 0 as kernel
is loaded and the card will indicate falsely that it doesn't support
UHS modes in kernel.
To prevent this, disable UHS mode support in U-boot so that the S18A
bit remains available and the card is able to boot in a UHS mode
in kernel.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>