configs: dra7xx_evm: add support for DRA71-lcard
This enabled booting a DRA71-lcard with the same binary as the other DRA7
based platforms.
After adding the dtb for the dra71-lcard, we need to increase the size of
the area where the compressed DTBs are uncompressed.

Note: The log messages of the SPL are not visible because the DTS used by
the SPL is by default the SPL of the DRA7-evm (console on UART0).

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>