dts: dra7x: Disable USB1 on all evms
On all the EVMs featuring a SOC of the DRA7 family, the USB1 port is used
as a device for DFU. This port is managed by the platform code and must not
be advertised to the DWC3 DM driver.
This will be changed when/if support for the device mode is added to the
dwc3-omap driver.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>